+++ languages/common/src/main/java/org/jetbrains/iren/VariableNamesContributor.java
@@ -20,7 +20,7 @@ public interface VariableNamesContributor {
      * @param predictionList container which contains all predictions.
      * @return priority of contribution
      */
-    int contribute(@NotNull PsiNameIdentifierOwner variable, @NotNull List<VarNamePrediction> predictionList);
+    double contribute(@NotNull PsiNameIdentifierOwner variable, @NotNull List<VarNamePrediction> predictionList);
 
     /**
      * Get conditional probability of variable name.
@@ -28,7 +28,7 @@ public interface VariableNamesContributor {
      * @param variable some variable.
      * @return pair of probability and model priority.
      */
-    @NotNull Pair<Double, Integer> getProbability(@NotNull PsiNameIdentifierOwner variable);
+    @NotNull Pair<Double, Double> getProbability(@NotNull PsiNameIdentifierOwner variable);
 
     @NotNull ModelType getModelType();
 }
